task: Implement isActive() into Level.js
isActive successfully checks if click target is being activated by
another pipe.
However, it doesn't de-activate it when it is no longer connected.

A solution could be an algorithm which pushes every pipe from beginning
to end in order to an array, and if a pipe isn't connected to the pipe
before it, it would be disconnected along with the pipes after it.